+++ back-end/src/main/java/admin/gpu/GpuRepository.java
@@ -1,4 +1,7 @@
 package admin.gpu;
 
+import org.springframework.stereotype.Repository;
+
+@Repository
 public class GpuRepository {
 }
