+++ back-end/src/main/java/admin/gpu/dto/WaitingJobDto.java
@@ -1,4 +1,4 @@
-package admin.gpu;
+package admin.gpu.dto;
 
 public class WaitingJobDto {
     private final String name;
